Add purchase pages
Add pages to facilitate purchases with the flathub-authenticator
web flow. The landing page takes the query parameters or stored pending
transaction, and it redirects to the login or checkout pages or to
flathub-authenticator. The checkout page is incomplete; this is where
Stripe payments go, and when that's done it should redirect to
/purchase.

Also modified the login flow so that if there's a pending transaction,
it redirects to /purchase after login rather than /userpage.