Direct user back to where they were upon login
Since login directs the user away from the site, we need to use local
storage to determine where to send them on return.

However, by sending them to the login page there's no guarantee that
they will go through with login. We don't want to pre-emptively set a
value which will later falsely be used if they return to login.

If that reasoning is unclear, also consider the scenario where a user
has two login windows open as a result of trying to visit different
protected pages. Whichever window they proceed with should direct them
back to where they were in that window, not the other window.

Using a URL query string means the local storage value can be set only
immediately before redirect away from the site, when we know they are
performing login. It can then be used from local storage on return.

If a value does remain in local storage, we prefer the query string
at time of redirect which will overwrite it.