feat(frontend): add back button during checkout
User may want to return to card selection if they change their mind or
accidentally click the button to use a new card.

I first tried this as a button in the Checkout module itself, below the
current stage of the flow. However, it feels too disconnected from the
current stage shown (unlike the cancel button which is independent of
the flow).

closes #134