fix: remove bad status console error if logged out
A status 204 seems more appropriate to tell the client the request was
successful, but there is no content to return (i.e. no user data when
logged out).

If we explicitly want to discourage use of this endpoint while logged
out we could leave it as 403 and instead work in some client-side user
context persistence with local storage. That's probably not worth the
hassle at this time.

fixes #296