Use long format for date(time) strings
Currently all regional variations of any given language are ignored:
UK and US English are treated as one language.

Sadly the numerical date format is different between the US (m/d/y) and
most of the rest of the anglosphere (d/m/y), and ambiguously so.

Use the "long" format for formatted dates. 'June 5, 2022' may not be the
way yesterday's date would be written in the UK, but unlike '6/5/2022'
it doesn't read as '6th May 2022'.

Fixes #329