feat: constrain currency input and add feedback
This turned out to be more complex than anticipated to enable a good UX.

I have implemented feedback following these guidelines:

- On input blur any feedback is shown. This means the user isn't
  bothered by the input preemptively assuming what they are entering is
  invalid before they finish.
- On input change any existing feedback is removed. The means the user
  instanly knows when they have fixed the problem.

The complexity really arises from the facts that:

- The rest of the form also needs to respond to invalid values in order
  to prevent submission (which can occur before the blur event)
- The string representation can acceptably be an invalid number
  during input
- The Flathub donation form needs to also be able to set the input value
  from outside of the component (so it must remain a controlled
  component)

In order to accomodate the various component needs the state is now an
object with a settled and live value, such that logic can be driven
based on the value after blur or during changes.

My changes also nicely encapsulate the string representation within the
input component so users of it no longer need to concern themselves with
the currency representation.