fix(frontend): redirect to stored page on login
The effect was running a second time when the stored value was
cleared, which results in default redirection to the user page.

I also updated the page to use the async hook and break the code up for
clairty. Could probably benefit from breaking logic out to components at
some point.