feat(frontend): support $0 minimum app vending
This required a rework of the components to pull the feedback behaviour
out of the CurrencyInput component so that custom conditions can be used
to drive a more specific message.

I opted to use wrapper components to avoid repeading logic, and
higher-order components seemed unsuitable since the behaviour is
customised for each component using them.