verification: Rewrite verification backend
The old verification system wasn't sufficient for our needs because it
didn't associate apps with a Flathub account. This commit mostly
rewrites the verification system to support that. The old
UserVerifiedApp table, which was only for login_provider verifications,
is dropped. All existing verifications (I'm only aware of one) need to be
redone.

The new verification system centers around the new AppVerification
table, which records all details associated with attempts to verify an
app. Website verification uses a random token stored in this table, and
is (currently) only checked once at the time of verification. This could
be updated to re-check periodically.

An app may only have one active verification at a time.

Manual verification now uses the table as well, instead of the JSON file
and redis. There is currently no endpoint to change this; it requires
inserting a row into the database. The benefit is less complexity since
all verification information is in the same table. A future project
could be to add "admin accounts" with an endpoint to verify/unverify
apps.