Always register vending api
Before this we needed to always set the stipe variables when we wanted to run codegen for the frontend, which was very annoying. This always registers the api, but stipe will likly fail, when running - which should be fine for testing.

This behavior also fooled local development, as developer portal management pages were not accessible due to these apis missing completly.