moderation: Skip generating moderation requests for runtimes
These come from trusted vendors and we don't want to block them in
the moderation queue / reviewer availability.

Any further apps or extensions (or refs in general) also coming from
these projects should not be skipped as there is a possibility that they
get dropped in a future runtime major release and moved to third-party.