moderation: Tag the last committer in rejection issue
This helps in cases where the repo has no watchers or the maintainer
is not subscribed

Assume the last committer of the default branch to be responsible for
the change. This should be the case for most of the things being
moderated ie. apps. This would fail for beta or extensions when
branch is not default but those are either rare or not moderated or
when multiple commits are merged.