apps: Improve subcategory indexing
Previously subcategories were calculated by removing the first
category from the category data supplied by the app and adding the rest
as subcategories. This approach leads to a lot of "junk" data in
subcategories for example, the index would have `2DGraphics` as
a subcategory of the main category `AudioVideo` or appstream's
low quality categories.

This fixes that by defining a map between parent categories and
subcategories obtained from the spec and adding only valid subcategories
for a given parent category as calculated from the map to the index